Round also converts to int if a number of digits wasn't specified
To be extra safe, medaltv's tests were run with Hatch

+++ yt_dlp/extractor/medaltv.py
@@ -102,7 +102,7 @@ class MedalTVIE(InfoExtractor):
             item_id = item_id or '%dp' % height
             if item_id not in item_url:
                 return
-            width = int(round(aspect_ratio * height))
+            width = round(aspect_ratio * height)
             container.append({
                 'url': item_url,
                 id_key: item_id,
